Move server.go into http.go
